Split out code from test
