+++ src/wrapped_radix_engine_toolkit/high_level/manifest_building/manifest_assets_transfers.rs
@@ -92,268 +92,6 @@ mod tests {
         let fung_1: ResourceAddress =
             ResourceAddress::sample_stokenet_gc_tokens();
 
-        #[derive(Clone, Debug, PartialEq, Eq, Hash, uniffi::Record)]
-        pub struct PerRecipientAssetTransfers {
-            pub address_of_sender: AccountAddress,
-            pub transfers: Vec<PerRecipientAssetTransfer>,
-        }
-
-        impl PerRecipientAssetTransfers {
-            pub fn new(
-                address_of_sender: AccountAddress,
-                transfers: impl IntoIterator<Item = PerRecipientAssetTransfer>,
-            ) -> Self {
-                Self {
-                    address_of_sender,
-                    transfers: transfers.into_iter().collect_vec(),
-                }
-            }
-        }
-
-        #[derive(Clone, Debug, PartialEq, Eq, Hash, uniffi::Record)]
-        pub struct PerRecipientAssetTransfer {
-            pub recipient: AssetsTransfersRecipient,
-            pub fungibles: Vec<PerRecipientFungibleTransfer>,
-            pub non_fungibles: Vec<PerRecipientNonFungiblesTransfer>,
-        }
-
-        #[derive(Clone, Debug, PartialEq, Eq, Hash, uniffi::Record)]
-        pub struct PerRecipientFungibleTransfer {
-            pub resource_address: ResourceAddress,
-            pub amount: Decimal192,
-            /// If `true` the `try_deposit_batch_or_abort` method will be used instead of `deposit`,
-            /// typically wallets sets this to try if and only if the recipient is a self-owned account
-            /// (`AssetsTransfersRecipient::MyOwnAccount`) controlled by a DeviceFactorSource thy have
-            /// access to and which third party deposit setting's `DepositRule` is `AcceptKnown` and
-            /// which resource is known (`resource_address` is owned or has been owned before).
-            use_try_deposit_or_abort: bool,
-            pub divisibility: Option<i32>,
-        }
-
-        #[derive(Clone, Debug, PartialEq, Eq, Hash, uniffi::Record)]
-        pub struct PerRecipientNonFungiblesTransfer {
-            pub resource_address: ResourceAddress,
-            /// If `true` the `try_deposit_batch_or_abort` method will be used instead of `deposit`,
-            /// typically wallets sets this to try if and only if the recipient is a self-owned account
-            /// (`AssetsTransfersRecipient::MyOwnAccount`) controlled by a DeviceFactorSource thy have
-            /// access to and which third party deposit setting's `DepositRule` is `AcceptKnown` and
-            /// which resource is known (`resource_address` is owned or has been owned before).
-            use_try_deposit_or_abort: bool,
-            local_ids: Vec<NonFungibleLocalId>,
-        }
-
-        impl PerRecipientNonFungiblesTransfer {
-            pub fn new(
-                resource_address: impl Into<ResourceAddress>,
-                use_try_deposit_or_abort: bool,
-                local_ids: impl IntoIterator<Item = NonFungibleLocalId>,
-            ) -> Self {
-                Self {
-                    resource_address: resource_address.into(),
-                    use_try_deposit_or_abort,
-                    local_ids: local_ids.into_iter().collect_vec(),
-                }
-            }
-        }
-
-        impl PerRecipientFungibleTransfer {
-            pub fn new(
-                resource_address: ResourceAddress,
-                amount: impl Into<Decimal192>,
-                use_try_deposit_or_abort: bool,
-                divisibility: impl Into<Option<i32>>,
-            ) -> Self {
-                Self {
-                    resource_address,
-                    amount: amount.into(),
-                    use_try_deposit_or_abort,
-                    divisibility: divisibility.into(),
-                }
-            }
-        }
-
-        impl PerRecipientAssetTransfer {
-            pub fn new(
-                recipient: impl Into<AssetsTransfersRecipient>,
-                fungibles: impl IntoIterator<Item = PerRecipientFungibleTransfer>,
-                non_fungibles: impl IntoIterator<
-                    Item = PerRecipientNonFungiblesTransfer,
-                >,
-            ) -> Self {
-                Self {
-                    recipient: recipient.into(),
-                    fungibles: fungibles.into_iter().collect_vec(),
-                    non_fungibles: non_fungibles.into_iter().collect_vec(),
-                }
-            }
-        }
-
-        impl From<(&AssetsTransfersRecipient, PerRecipientFungibleTransfer)>
-            for PerAssetTransfersOfFungibleResource
-        {
-            fn from(
-                value: (
-                    &AssetsTransfersRecipient,
-                    PerRecipientFungibleTransfer,
-                ),
-            ) -> Self {
-                let (recipient, fungible_with_amount) = value;
-                Self::new(
-                    PerAssetFungibleResource::new(
-                        fungible_with_amount.resource_address,
-                        fungible_with_amount.divisibility,
-                    ),
-                    [PerAssetFungibleTransfer::new(
-                        recipient.clone(),
-                        fungible_with_amount.use_try_deposit_or_abort,
-                        fungible_with_amount.amount,
-                    )],
-                )
-            }
-        }
-
-        impl From<(&AssetsTransfersRecipient, PerRecipientNonFungiblesTransfer)>
-            for PerAssetTransfersOfNonFungibleResource
-        {
-            fn from(
-                value: (
-                    &AssetsTransfersRecipient,
-                    PerRecipientNonFungiblesTransfer,
-                ),
-            ) -> Self {
-                let (recipient, non_fungibles) = value;
-
-                PerAssetTransfersOfNonFungibleResource::new(
-                    non_fungibles.resource_address.clone(),
-                    [PerAssetNonFungibleTransfer::new(
-                        recipient.clone(),
-                        non_fungibles.use_try_deposit_or_abort,
-                        non_fungibles.local_ids,
-                    )],
-                )
-            }
-        }
-
-        impl From<(&AssetsTransfersRecipient, PerRecipientFungibleTransfer)>
-            for PerAssetFungibleTransfer
-        {
-            fn from(
-                value: (
-                    &AssetsTransfersRecipient,
-                    PerRecipientFungibleTransfer,
-                ),
-            ) -> Self {
-                let (recipient, transfer) = value;
-                Self::new(
-                    recipient.clone(),
-                    transfer.use_try_deposit_or_abort,
-                    transfer.amount,
-                )
-            }
-        }
-
-        impl From<(&AssetsTransfersRecipient, PerRecipientNonFungiblesTransfer)>
-            for PerAssetNonFungibleTransfer
-        {
-            fn from(
-                value: (
-                    &AssetsTransfersRecipient,
-                    PerRecipientNonFungiblesTransfer,
-                ),
-            ) -> Self {
-                let (recipient, non_fungibles) = value;
-                Self::new(
-                    recipient.clone(),
-                    non_fungibles.use_try_deposit_or_abort,
-                    non_fungibles.local_ids,
-                )
-            }
-        }
-
-        impl PerAssetTransfersOfFungibleResource {
-            fn expanded(
-                &mut self,
-                transfer: impl Into<PerAssetFungibleTransfer>,
-            ) {
-                self.transfers.push(transfer.into());
-            }
-        }
-
-        impl PerAssetTransfersOfNonFungibleResource {
-            fn expanded(
-                &mut self,
-                transfer: impl Into<PerAssetNonFungibleTransfer>,
-            ) {
-                self.transfers.push(transfer.into());
-            }
-        }
-
-        impl PerRecipientAssetTransfers {
-            pub fn transpose(&self) -> PerAssetTransfers {
-                let mut per_asset_fungibles = IndexMap::<
-                    ResourceAddress,
-                    PerAssetTransfersOfFungibleResource,
-                >::new();
-
-                let mut per_asset_non_fungibles = IndexMap::<
-                    ResourceAddress,
-                    PerAssetTransfersOfNonFungibleResource,
-                >::new();
-
-                self.transfers.clone().into_iter().for_each(|t| {
-                    let x = t.clone();
-                    let recipient = &x.recipient;
-
-                    x.fungibles.clone().into_iter().for_each(|y| {
-                        if let Some(existing_transfers) =
-                            per_asset_fungibles.get_mut(&y.resource_address)
-                        {
-                            existing_transfers.expanded((recipient, y.clone()));
-                        } else {
-                            per_asset_fungibles.insert(
-                                y.resource_address.clone(),
-                                PerAssetTransfersOfFungibleResource::new(
-                                    PerAssetFungibleResource::new(
-                                        y.clone().resource_address,
-                                        y.clone().divisibility,
-                                    ),
-                                    [PerAssetFungibleTransfer::from((
-                                        recipient,
-                                        y.clone(),
-                                    ))],
-                                ),
-                            );
-                        }
-                    });
-
-                    x.non_fungibles.clone().into_iter().for_each(|y| {
-                        if let Some(existing_transfers) =
-                            per_asset_non_fungibles.get_mut(&y.resource_address)
-                        {
-                            existing_transfers.expanded((recipient, y.clone()));
-                        } else {
-                            per_asset_non_fungibles.insert(
-                                y.resource_address.clone(),
-                                PerAssetTransfersOfNonFungibleResource::new(
-                                    y.clone().resource_address,
-                                    [PerAssetNonFungibleTransfer::from((
-                                        recipient,
-                                        y.clone(),
-                                    ))],
-                                ),
-                            );
-                        }
-                    });
-                });
-
-                PerAssetTransfers::new(
-                    self.address_of_sender.clone(),
-                    per_asset_fungibles.values().cloned(),
-                    per_asset_non_fungibles.values().cloned(),
-                )
-            }
-        }
-
         let per_recipient_transfers = PerRecipientAssetTransfers::new(
             sender.clone(),
             [
