+++ src/wrapped_radix_engine_toolkit/high_level/sargon_specific_types/assets_transfers/mod.rs
@@ -1,5 +1,9 @@
 mod assets_transfers_recipient;
 mod per_asset;
+mod per_recipient;
+mod transfer_types;
 
 pub use assets_transfers_recipient::*;
 pub use per_asset::*;
+pub use per_recipient::*;
+pub use transfer_types::*;
