+++ src/wrapped_radix_engine_toolkit/high_level/sargon_specific_types/assets_transfers/per_asset/mod.rs
@@ -1,11 +1,9 @@
 mod per_asset_fungible_resource;
-mod per_asset_transfer_types;
 mod per_asset_transfers;
 mod per_asset_transfers_of_fungible_resource;
 mod per_asset_transfers_of_non_fungible_resource;
 
 pub use per_asset_fungible_resource::*;
-pub use per_asset_transfer_types::*;
 pub use per_asset_transfers::*;
 pub use per_asset_transfers_of_fungible_resource::*;
 pub use per_asset_transfers_of_non_fungible_resource::*;
