+++ src/wrapped_radix_engine_toolkit/high_level/sargon_specific_types/assets_transfers/per_asset/per_asset_transfers_of_fungible_resource.rs
@@ -7,6 +7,15 @@ pub struct PerAssetTransfersOfFungibleResource {
 }
 
 impl PerAssetTransfersOfFungibleResource {
+    pub(crate) fn expanded(
+        &mut self,
+        transfer: impl Into<PerAssetFungibleTransfer>,
+    ) {
+        self.transfers.push(transfer.into());
+    }
+}
+
+impl PerAssetTransfersOfFungibleResource {
     pub fn new(
         resource: PerAssetFungibleResource,
         transfers: impl IntoIterator<Item = PerAssetFungibleTransfer>,
@@ -29,6 +38,27 @@ impl PerAssetTransfersOfFungibleResource {
         let rounded = amount.round(self.resource.divisibility);
 
         rounded.into()
+    }
+}
+
+impl From<(&AssetsTransfersRecipient, PerRecipientFungibleTransfer)>
+    for PerAssetTransfersOfFungibleResource
+{
+    fn from(
+        value: (&AssetsTransfersRecipient, PerRecipientFungibleTransfer),
+    ) -> Self {
+        let (recipient, fungible_with_amount) = value;
+        Self::new(
+            PerAssetFungibleResource::new(
+                fungible_with_amount.resource_address,
+                fungible_with_amount.divisibility,
+            ),
+            [PerAssetFungibleTransfer::new(
+                recipient.clone(),
+                fungible_with_amount.use_try_deposit_or_abort,
+                fungible_with_amount.amount,
+            )],
+        )
     }
 }
