+++ src/wrapped_radix_engine_toolkit/high_level/sargon_specific_types/assets_transfers/per_asset/per_asset_transfers_of_non_fungible_resource.rs
@@ -19,6 +19,34 @@ impl PerAssetTransfersOfNonFungibleResource {
 }
 
 impl PerAssetTransfersOfNonFungibleResource {
+    pub(crate) fn expanded(
+        &mut self,
+        transfer: impl Into<PerAssetNonFungibleTransfer>,
+    ) {
+        self.transfers.push(transfer.into());
+    }
+}
+
+impl From<(&AssetsTransfersRecipient, PerRecipientNonFungiblesTransfer)>
+    for PerAssetTransfersOfNonFungibleResource
+{
+    fn from(
+        value: (&AssetsTransfersRecipient, PerRecipientNonFungiblesTransfer),
+    ) -> Self {
+        let (recipient, non_fungibles) = value;
+
+        PerAssetTransfersOfNonFungibleResource::new(
+            non_fungibles.resource_address.clone(),
+            [PerAssetNonFungibleTransfer::new(
+                recipient.clone(),
+                non_fungibles.use_try_deposit_or_abort,
+                non_fungibles.local_ids,
+            )],
+        )
+    }
+}
+
+impl PerAssetTransfersOfNonFungibleResource {
     pub fn all_ids(&self) -> Vec<ScryptoNonFungibleLocalId> {
         self.transfers
             .clone()
